+++ apps/ngx-localstorage-demo/src/app/components/directive/directive.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ngx-localstorage-directive',
+  templateUrl: './directive.component.html',
+  styles: []
+})
+export class DirectiveComponent {
+
+  defaultFalsyTransformer: () => any = () => false;
+}