+++ apps/ngx-localstorage-demo/tailwind.config.js
@@ -1,0 +1,17 @@
+const { createGlobPatternsForDependencies } = require("@nrwl/angular/tailwind");
+const { join } = require("path");
+
+module.exports = {
+  corePlugins: {
+    preflight: false,
+  },
+  content: [
+    join(__dirname, "src/**/!(*.stories|*.spec).{ts,html}"),
+    ...createGlobPatternsForDependencies(__dirname),
+  ],
+  darkMode: 'media',
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};