+++ libs/ngx-localstorage/src/lib/services/ngx-localstorage.service.spec.ts
@@ -2,8 +2,6 @@ import { TestBed, inject } from '@angular/core/testing';
 
 import { LocalStorageService } from './ngx-localstorage.service';
 import { NGX_LOCAL_STORAGE_CONFIG } from '../tokens/storage-config';
-import { DefaultSerializer } from '../classes/default-serializer';
-import { NGX_LOCAL_STORAGE_SERIALIZER } from '../tokens/storage-serializer';
 
 describe('LocalStorageService', () => {
   beforeEach(() => {
@@ -14,10 +12,6 @@ describe('LocalStorageService', () => {
           useValue: {
             prefix: 'ngx-localstorage'
           }
-        },
-        {
-          provide: NGX_LOCAL_STORAGE_SERIALIZER,
-          useClass: DefaultSerializer
         }
       ]
     });
@@ -70,6 +64,22 @@ describe('LocalStorageService', () => {
 
       const key = service.getKey(0);
       expect(key).toBe(`${service.config.prefix}_${prefixlessKey}`)
+    }
+  ));
+
+  it('should add entries with passed options', inject(
+    [LocalStorageService],
+    (service: LocalStorageService) => {
+
+      const prefixlessKey = 'entry';
+      const perCallPrefix = 'per-call-prefix';
+
+      service.set(prefixlessKey, 'value', {
+        prefix: perCallPrefix
+      });
+
+      const key = service.getKey(0);
+      expect(key).toBe(`${perCallPrefix}_${prefixlessKey}`)
     }
   ));